InfluencerForm/requirements.txt
Flask==3.1.0
gspread==6.2.1
oauth2client==4.1.3
gunicorn==23.0.0
python-dotenv==1.1.1

